activate the logging of executed SQL statements
SQL category to DEBUG (or the corresponding log level of your log
framework). Hibernate will write the SQL statements without any
information about the values bound to the JDBC parameters or extracted
from the query result. If you also want to get this information, you
have to set the log level of org.hibernate.type.descriptor.sql to DEBUG
(see the development recommendation for an example).